add actClamp as ECM, combine BON and ON into one ON since its one algorithm.
