+++ online-norm/tensorflow/README.md
@@ -82,3 +82,15 @@ or
 ```bash
 nosetests
 ```
+
+## change log
+
+### 2020-03-25
+
+#### Added
+
+- Added [activation clamping](LinkToActClampPaper) as an error checking mechanism (ecm) and make it the defualt ecm.
+
+#### Changed
+
+- Combine BatchedOnlineNorm and OnlineNorm into one class since its one algorithm.
