+++ online-norm/tensorflow/online_norm_tf/__init__.py
@@ -1,7 +1,6 @@
 """
-Released under BSD 3-Clause License, 
+Released under BSD 3-Clause License,
 Copyright (c) 2019 Cerebras Systems Inc.
 All rights reserved.
 """
 from .online_norm import OnlineNorm, online_norm
-from .batch_online_norm import BatchOnlineNorm, batch_online_norm
